Создала DOM-элементы, соответствующие фотографиям
