+++ js/main.js
@@ -1,4 +1,7 @@
 import './utils.js';
+import './picture.js';
 import { getRandomInt } from './utils.js';
+import {resize} from './picture.js';
 
 getRandomInt(1, 2);
+resize();
